Added screen updates and timeout
Still tracking down a bug that causes a hang intermittenly. Have tried
implementing a watchdog timer (2sec) to see if that at least gets the
reset working, but it's still not clear where the hang is coming from. The
real time clock continues running and outputting 32.768kHz signal, so it's
not an issue with the power supply. Button2 is currently co-opted as an
output in this version to help with monitoring on the scope ( bitSet(PIND,
3) peppered throughout)