Adding flag to declare full OLED screen redraws
Set screenChange = true to force the OLED screen to redraw the channel
labels, or false to only update the Hall effect values (faster).
Interestingly, the optiboot bootloader appears to be jumping to the main
program too fast for my home ftdi programmer, so I have to hit the reset
button just prior to attempting a new firmware upload over serial.