Experimenting with only taking readings on save interval
Test case where Hall sensors are only read when the takeSamples flag is
true and either the OLED screen is on or the writeData flag is true.
This has introduced a symptom where the current draw is not returning to
normal after a SD write until another round of Hall sensor readings are
taken.