Adding delay + shift register clear after writeToSD
Was having a problem with high current draw immediately after writes to
the SD card. It appears the shift register (which shares the SPI bus
with the SD card) wasn't happy with whatever signals were being sent to
the SD card and was occasionally ending up in some kind of high current
draw state. Adding a 10ms delay and an additional clear() call to the
shift registers after the writeToSD call appears to have removed the
high current draw issue.