Trying to track down hang cause
The device is occasionally reseting, possibly due to watchdog timeout, for unknown reasons. Somethings this happens within 1 minute of a reboot. In other cases, the device appears to hang with the green LED flashing quickly, permanently. Changed reboot flash from green to red to see if that is the source of the permanent green-flash hang.